PHP expose_php Information Disclosure
The configuration of PHP on the remote host allows disclosure of
sensitive information.

Description
The PHP install on the remote server is configured in a way that allows
disclosure of potentially sensitive information to an attacker through a
special URL. Such a URL triggers an Easter egg built into PHP itself.

Other such Easter eggs likely exist, but Nessus has not checked for
them.

Solution
In the PHP configuration file, php.ini, set the value for 'expose_php'
to 'Off' to disable this behavior. Restart the web server daemon to put
this change into effect.

Turn off expose_php in php.ini

This fixes issue MON-11321.

Signed-off-by: Ramesh T ramesht@op5.com