[LDAP] Add option to find groups by using memberof
This commits adds the ability to search for group membership by doing a
query on the user and use the memberOf attribute. This is instead of
querying for all group and looking if the user is a member of that
group.

The option needs to be enable by configuration.

This is part of: MON-11271

Signed-off-by: Jacob Hansen <jhansen@op5.com>