ninja: Add new comment in extinfo page
When a new comment is added in extinfo page of host or service,
 it is not visible immidiately. However after couple of seconds
followed by page refresh brings back the added comment.
This issue is only visible in lmd integrations to monitor.
The reason being a newly fired command to naemon is not immidiatle
 reflected. This fix addresses this issue.

MON-11281

Signed-off by: sbhapur@op5.com