Bugfix: Table header shown twice in listviews with Nachos banners
Earlier, when Nachos had a banner to show, the listviews would get 2 table
headers. The listviews had 2 table headers earlier as well, only that it
wasn't as clear. The table header was cloned to be fixed on page when table
was scrolled, and the original header lied under the clone. The Nachos banner
then pushed down and showed the original table header.
This update removes the original header when there's a Nachos banner. But
the clone will remain.

This fixes MON-11123

Signed-off-by: Elin Linder <elinder@op5.com>