ninja: Change default limit option
The default value of a limit filter is currently set to 0 in ninja.
Which is accepted by livestatus but not by LMD as a value of 0 is
getting ignored. For a huge configuration LMD's response is very large
in this case which violates default php limit of 128M.
This fix addresses this issue by modifying default limit from 0 to a
valid value for lmd which is 100 in this case.

MON-11384

Signed-off by:sbhapur@op5.com