Correctly verify naemon conf at backup
The backup function did not correctly run the naemon binary as the
monitor user. This resulted in the configuration not being able to be
verified correctly and the backup would never be made.

Futhermore the error message was not presented to the user due to
accessing the JSON data returned from the verify function.

Signed-off-by: Jacob Hansen <jhansen@op5.com>