Do not attempt to URL validate empty/false setting
The return value of fetch_page_setting might be false. However we would
still attempt to do a foreach loop over this value. This would cause the
ninja log to fil up quickly on many systems, and potentially break some
settings.

This commit ensures that we just return false from validate_uri_string
in case the setting is false.

This fixes: MON-11475

Signed-off-by: Jacob Hansen <jhansen@op5.com>