Backup: Handle all cases of fail json
Commit: 3d554db2ea2 handled one case of a incorrect handling of a JSON
structure. This ensured that errors during Naemon verification would correctly
be presented in the UI. However the are a few other similar cases in the
backup.js that would have the same issue.

These are handled in this commit, and should ensure all failures would
display correct information to the user.

Signed-off-by: Jacob Hansen <jhansen@op5.com>