Use specific version of parallel ruby gem
Previously the parallel ruby gem was not listed amongst the gem to be
installed. It was installed as a dependency to parallel_tests and thus
the latest available version was always installed. When parallel version
1.14 was released it required ruby version 2.2 or higher. We only have
ruby 2.0 and thus our builds fail.

With this commit we specify that parallel version 1.13 is going to be
installed. 1.13 is the latest version that supports ruby 2.0.

Signed-off-by: Daniel Nilsson <dnilsson@op5.com>