Ninja : Edit User Menu only save the last group edited
Earlier when we create 2 or more groups under 'Assign group rights' and
try to add or edit the menu ('Edit user menu' under "My account"), it
was removing previously added menu of other groups.

we changed the update setting so that it save the
previous config as it is.

This fixes issue MON-8286

Signed-off by Devesh Thakur dthakur@op5.com