New product image and changes in menu and on about page
We've reworked the brand function that fetches the logo since there's supposed
to be different product images on login page and in menu. If menu calls the
brand function it will also send a parameter to show the special product image.

When you click on the product image you will see another menu. Previously there
where icons in this menu, but they are removed.
The About page has also got the new product image, and is now aligned to the
left.

This is a part of MON-11626

Signed-off-by: Elin Linder <elinder@op5.com>