Apply recurring downtimes on Sundays
NinjaDateTime uses ISO 8691 to express dates and time, with week days
numbered as 1-7, 7 being Sunday. While the input data uses the North
American/Islamic system which numbers week days as 0-6, 0 being Sunday.
This caused recurring downtimes scheduled on a Sunday to fail, while all
other week days worked correctly. With this commit the function which
returns the day of the week is updated to use the 0-6 format.

Fixes: MON-11684

Signed-off-by: Erik Sjöström <esjostrom@op5.com>