Removing password-helper height declaration
When creating the initial admin account in Monitor, this style
is used for a message telling the user that admin credentials
cannot be recovered.

This element should not have a static height, as this will cause
a message that is longer to overflow outside of the div.

Instead, no height should mean that the box to expands downwards
as more text is added.

This fixes: MON-11748
"Text overflow when creating admin account on fresh install"