Make sure LMD is running after installation
During the installation of ninja, in the post section, we call a  php script,
migrate_tac_hostperf_to_listview.php and. However on EL6, for some reason LMD
is not running at this point, which is causing the scripts to error out due to
the livestatus socket not being present.

This commit ensures that the LMD is running before these scripts are
executed.

Signed-off-by: Jacob Hansen <jhansen@op5.com>