Use DEL_DOWNTIME_BY_HOST_NAME to delete downtimes
To delete a downtime ninja would prior to this commit use the external
command DEL_HOST_DOWNTIME or DEL_SVC_DOWNTIME. However this would fail
to remove downtimes for peers and pollers iff the downtime was not yet
started.

This happens for two reasons:
1) Merlin cannot safely propagate DEL_HOST/SVC_DOWNTIME to other nodes
as this external command requires a downtime_id argument which is not
guaranteed to be the same on all nodes.

2) Instead Merlin relies on deleting downtimes when a
NEBATTR_DOWNTIME_STOP_CANCELLED event happens, which is propagated
across to all relevant nodes. However this event only occurs if a
downtime is deleted AFTER the downtime start time, and as a result it
would therefore be missed.

Instead we will use DEL_DOWNTIME_BY_HOST_NAME which we can safely
propagate to all relevant nodes, and it will correctly delete the
downtime. Furthermore this makes downtime deletion behaviour more
cosistent in case there are several downtimes for the same object, with
the same starttime and comment.

This fixes: MON-9109.

Signed-off-by: Jacob Hansen <jhansen@itrsgroup.com>