Adapt report header for logger reports
Stop try to print time zone and state types for logger reports.

When doing a "log archive search" from the Logger listview, a report of
type "logger" is created, which shares report header code with other
report types. But for a logger report, some data, like the
`report_timezone` added in MON-11209 is not available which causes a
call to `timezone_utc_no()` to crash.

This fixes MON-11763.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>