Re-write cron function to get periodic schedules
This is a major re-write of the get_period_schedules() function, that
had some faults in it, causing report to not be sent or sent on the
wrong days.

This implementation should be more robust, since it doesn't rely on that
the system clock is exactly the same as the scheduled reports selected
report time.

However, this change doesn't claim to be a perfect solution to
scheduling reports.

Daily scheduled reports
-----------------------
With this change daily scheduled reports are generated as soon as
current time is beyond one report period after the report was previously
sent.

Example schedule:
- Daily report using timezone UTC
- Report time is 12:00 (same timezone as report)
- Scheduled to be sent every 2nd day

If the report was last sent anytime at 2019-10-15, it will be sent
first thing when the code runs after `2019-10-17 12:00 UTC`.

Weekly scheduled reports
------------------------
This change should fix some issues with the weekly scheduled reports,
like that they were sometimes sent on the wrong day.

Monthly scheduled reports
-------------------------
This change should fix some issues with the monthly scheduled reports,
for example that it wouldn't be sent if it for some reason missed a
month.

This fixes MON-11823.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>