Datapicker broken in event log
If you click on the calendar picker from event log's report, weekdays
don't looks correctly. Also, a weird character is next to the year.

With this commit, those errors will be fixed. Each day of the week is
setted in the good way and the weird character has been removed.

This fixes MONUI-58

Signed-off-by: Adrián Villalba <avillalba@itrsgroup.com>