Fix weekly scheduled reports on multiple days
This fixes a bug where if a report is scheduled weekly on multiple days,
it was sent every time the code ran on the subsequent scheduled days,
since `last_sent` wasn't updated. With this change, `last_sent` updates
every time a weekly schedule is sending the report.

And improve a couple of readability things.

This fixes MON-11823.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>