Fix report schedules for reports with missing timezone
With this patch we will not assume that there is a 'report_timezone'
report option stored, which is the case for reports that were saved
before the 'report_timezone' option was introduced.

Log to Ninja log on notice level (not enabled by default) when sending
reports, so that it can be tracked easier whether we're trying to send
reports when scheduled or not.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>

This fixes MON-11912.