Clarify description text for commands
These descriptions are visible when acknowledging the state of
an object in the UI, through the extinfo page or via the drop-down.

1) When using a checkbox in a UI, it's implicit that leaving that
box checked is a "yes" answer to the text provided, so there is no
need to say "If you want ... check this checkbox". This change
makes the descriptions shorter and hopefully more clear.

2) On customer request, we have added a note that if the object is
in a scheduled downtime, the acknowledgement notification will be
sent regardless. This is maybe not intuitive or obvious, but it's
the way it is so the text should probably underline this fact.

This fixes MONUI-155