Fix permissions for limited_edit group
Add the permissions "host_template_view_all" and
"service_template_view_all" to the limited_edit group in the default
config group right configuration. This fixes an error where a user in
the limited_edit group cannot edit an object for which he/she is a
contact for.

This commit also;
- Sorts list of permissions within each group.
- Remove duplicated permissions.

This fixes MON-11987.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>