Fix delete downtime by hostname action in GUI
Following adding the option to delete downtimes via the API, an unwanted
side effect occured, where we got a new "Delete Downtime" option in
listviews. This did not work, causing confusing.

It appears as though it is not possible to hide this action from the
listviews, without affecting the API. This commit therefore fixes the
view so that the command can be sent successfully via the GUI.

This fixes: MON-12088

Signed-off-by: Jacob Hansen <jhansen@op5.com>