Fix double quicklins when indexes are out of order
In the previous commit: 925b87b duplicate quicklinks was fixed.

However in some cases the fix would not work. The reason for this was
that array_unique would leave index gaps in an indexed. That is you
could end up with an array like:

```
[0] => ...
[2] => ...
```

When then encoding the array with `json_encode`, the we would get an
indexed array like so:
`{"0":{quicklink}, "2":{quicklink}`
instead of the desired format:
`[{quicklink},{quicklink}]`

Signed-off-by: Jacob Hansen <jhansen@op5.com>