Simplify downtime command generation
The code which generated the downtime commands used the php function
str_replace to build the command, this meant that all the words which
were to be replaced were in essence reserved. This was found in the bug
MON-12387 where a hostgroup which contained 'start' in its name had that
word replaced by the start-time of the downtime leading to Naemon not
being able to execute the command since the hostgroup had its name
changed.

With this commit the usage of str_replace is replaced by a simple
concatenation of values.

Fixes: MON-12387

Signed-off-by: Erik Sjöström <esjostrom@op5.com>