Fix PNG path to avoid 404 errors
I believe that any normal GUI user of Ninja will always have a theme
set, whether that is "default" or something else. If that is the case,
this user will have no problem accessing "bg.png" (whatever it's used
for) because images/bg.png is a relative path that makes sense inside
the default, classic, pink_n_fluffy dirs etc.

However, this relative path does not make sense in the
application/views/css directory which this file is in, because there is
no "images" directory here.

This means that anything using this CSS file will get a 404 error, as it
will try to request an image file that does not exist. So why is this a
problem?

It seems that one of the components that actually use this "non-theme"
is the HTML that is rendered to go into a report via wkhtmltopdf, so you
can reproduce this 404 error by trying the following:

- Generate a PDF report from the GUI by clicking "As PDF" in the top
right corner.
- Grep for 404 in your /var/log/httpd/ssl_access_log and you will see
something like the following:

"GET /monitor/application/views/css/images/bg.png HTTP/1.1" 404 241

Note that this is trying application/views/css/images, which does not
exist.

So again, why is this a problem? It's a problem because it causes
wkhtmltopdf to exit with a non-zero return code, which is visible while
debug logging in ninja.log:

The first line tells you the command that was run and the error code, if
any:

"Pdf command /usr/bin/wkhtmltopdf (...) returned 2:"

The second line tells you what STDERR gave back:

stderr: Exit with code 2 due to http error: 404 Page not found

The report generation still works, but I don't think this is a reason to
accept constantly throwing an error code when there's no actual issue.

A reviewer should be able to confirm this patch fixes the issue by first
noting the above output every time a report is generated via clicking
"As PDF", and then see it disappear after this change.