Filter input/output on dashboard names
This filters input/output on dashboard names for dangerous javascript.

The filter on the input is not needed as ModSec filters that correctly,
but I have put it in as a failsafe if modsec is disabled.  If you
somehow managed to get a dangerous name in the Merlin database the
output filtering should avoid any javascript from executing

This fixes: MON-12468 which is part of the EMCA Pen tests (MON-12205)

Signed-off-by: Axel Bolle <abolle@itrsgroup.com>