Change strotime input to correctly render 'last week'
This works fine on mondays and tuesdays, but breaks for all other days
of the week, likely due to the 'midnight -1 sec' rescuing monday and
tuesday by actually making them sunday and monday, but the 'monday'
passed for time_end then becomes monday next week, causing the report to
show everything up until now.

This commit changes the overly verbose strtotime inputs to something
that hopefully more accurately should reflect "last week" regardless of
what day of the week it is.

For testing/reproduction, I suggest disabling NTP and manually changing
the OS time, then refreshing a report to see how it changes with the
days.

This fixes MON-12548 which has more information in Jira.