Fixed 'last week' time period
This commit fixes an issue where the 'last week' period would only work
correctly when run on Monday and Tuesday, but for any other day of the
week would display report data up to the current time.

Changing the input to strtotime() fixed this for all days except Sunday.
While investigating, we encountered a bug in PHP which parses strings
like "monday last week" incorrectly. A workaround has been added that
should only take effect on PHP versions < 5.6.29. More info in
MON-12548.

Additional tests were added, so instead of only Monday, we are now
testing all days of the week.