Update openid library with basic error handling
When PKCE is not enabled Keycloak does not update the session with
a codeVerifier which the openid library expects. With this commit we
add error handling to the library to check if the code is set before
trying to get it. As a consequence we can also remove some hacks
implemented in order to make it work.

Signed-off-by: Erik Sjöström <esjostrom@itrsgroup.com>