Complete AuthDriver_Keycloak API and create new users
This commit cleans up the AuthDriver_Keycloak class. We make a concious
decision to include Keycloak users in the local user data stored on
file. We do this to simplify the implementation of the other methods
expected in the AuthDriver API, which are also implemented with this
commit (basically copy pasted from the Default driver, with the
exception that we only look at Keycloak users).

With this commit we also create Keycloak users that does not already
exist as local users. This relies on Keycloak including the "grp" claim
that includes the group memberships the user should be created with.

This is part of MON-12654.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>