Use apcu instead of apc
The APC library has been superseded by a new library called APCu. There
exists a compatibility package called apcu-bc but it does not exist in
standard repos. Rather than trying to install the compatibility package
we try to update our use of the APC functions with the APCu equivalents.

This is part of MON-12644.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>