Use password_hash() instead of crypt()
Using the crypt() function without a salt argument is deprecated and
will be removed in PHP 8. Already in PHP 7 this raises a warning that is
treated as an error by PHPUnit.

This commit replaces our use of the crypt() function with
password_hash(), as recommended by PHP documentation. Supposedly
password_hash() is compatible with crypt() so hopefully it will not
cause any problems.

See also: https://www.php.net/manual/en/function.crypt.php

This is part of MON-12644.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>