Fix signature of Dashboard_Model::set_read_perm
The function signature of Dashboard_Model::set_read_perm did not match
that of the parent class, which is an error since PHP 7.something. This
commit changes the function signature to match the untyped signature in
the parent class, and adds some type checking for good measure.

This is part of MON-12647.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>