Use password_verify
With the update to PHP 7 we changed the method used to hash passwords,
from `crypt` to `password_hash`. While the documentation for
`password_hash` notes that it is compatible with `crypt`, this does not
mean they are identical. The hashed outputs are actually different
between the two methods. So we can no longer just compare the hashed
strings to validate our passwords, since that will not work for legacy
password created with `crypt`. The solution is to use `password_verify`
instead, that should be able to correctly verify hashed passwords
created with both `crypt` and `password_hash`.

This is part of MON-12586.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>