Prevent login from redirecting to unknown sites
The login page allows us to set a `uri` query parameter to an address
where the user will be redirected after a successful login (or directly
redirected if already logged in). This commit limits this feature to
only work for pages on the same site. Attempts at redirecting to a
different site will instead be redirected to the default start page.

This type of open redirects could be used in phishing attacks, so it's
safer to disallow it. There doesn't appear to be any use cases in
Monitor that would warrant a redirect outside the site, afaict.

If we wanted to be even more rigid we could probably implement this type
of check in the `redirect` helper function itself, but then the impact
becomes much broader and it might break some valid use case out there.
So i have opted to not do that.

This is part of MON-12841.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>