Update tooltip for listview column config
There was some confusion around the column config option `all`, and why
it didn't return more columns than the `default` option. This updates
the tooltip to instead refer to `default`, which also matches the
documentation.

It appears that `all` will return the same set of columns as `default`,
which are all those that are defined in the file
modules/monitoring/src/js/lsfilter_renderer_table.js.

This is part of MON-12901.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>