Remove option to change user theme
With this commit we remove the option for a user to change the theme for
the monitor UI and reports. We do this change not because theming is a
bad thing, but because we currently don't have the bandwidth to maintain
them and because some might find the pink & fluffy theme to be a tad bit
unprofessional.

Since we don't actually think theming is a bad idea I only removed the
ability for a user to change the theme and left the actual behind the
scenes code intact. This also has the benefit of greatly lessening the
work I have to do which I consider to be a big win.

Implements: MON-12935

Signed-off-by: Erik Sjöström <esjostrom@itrsgroup.com>