Fix host states links
The host states links on the extinfo page for hostgroups were broken.
They were trying to filter on 'host.groups' rather than just 'groups'.
This resulted in a lsfilter query like this:

    query: [hosts] host in "<hostgroup_name>" ...

Which results in a 500 error because 'host' is not a recognized field.

This commit changes the links to just 'groups', which makes them work as
expected. It does not address the 500 error, however, because the
efforts involved proved insurmountable by the author.

This is part of MON-12830.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>