Redirect / to /monitor/index.php
With the removal of the OP5 Portal in Monitor 9, we don't have any
landing page on the root path of the Monitor web server. This commit
adds a redirect from the the root web path to the Ninja site, which in
turn is redirected depending on whether there is an active session or
not. A new user browsing to the Monitor server will then typically come
to the Ninja login page instead.

This is part of MON-12622.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>