Add feature test for backup/restore config
Add a test case that does a backup & restore of configuration through
the UI.
This test would reproduce the problem in MON-12985, if monitor-http-api
were installed, which is not normally the case when this test suite
runs.
However, it should fail in case similar issues arise in the future,
where some file included in the backup is not writable by the
unprivileged user that runs restore (httpd process).

Part of MON-12985.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>