Ensure $controller->template is an object
In prior version of PHP, attempts to access a `null` variable as an
object, would cause PHP to create a new object from the default object:
`stdClass`. However in PHP7.4 this is no longer the cases and causes an
error:

```
Creating default object from empty value in
/opt/monitor/op5/ninja/application/hooks/load_js_bundles.php on line 11
```

In this commit, we first check whether the `template` object is set (i.e
not null), and if it's not, we create a new object from `stdClass`.
This is functionally equivalent to how the code works on earlier
versions of PHP.

Part of MON-12965

Signed-off-by: Jacob Hansen <jhansen@op5.com>