Fix saving default graph from PNP page
Graphs on the PNP page has a button that lets us save it as the "default
graph". (It is unknown to me what being the default graph actually does,
though.) This functionality was broken since a recent update to
PNP4Nagios that changed the order in which the source and view query
parameters appear in the URL, which is then fixed by this commit. (This
is obviously a very fragile way of getting the data in the first place,
but I'm not going to bother doing more than necessary at this point.)

This is part of MON-13020.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>