Replace hard coded URLs with values from config
Replace URLs to product documentation and support with values from a
Ninja configuration file.

Add generic mechanism for other packages to override config files by
putting a file with the same name in override.d/. This will be used by
op5-monitor package to write a product.php file with version specific
documentation URL.
Ninja still needs to provide the product.php file for it to work when
op5-monitor package is not installed (post-install tests for ninja and
various other packages).

Part of MON-13029.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>