Implement/fix default graph hover in listview
This commit will introduce a new behaviour for graphs.
When you now hover over the graph icon in a listview a lookup will be done
to see if the setting "default_graph" is set on that particular host. If
set, it will display the corresponding graph from pnp.

Also improved the way we save default_graph in the database. It is now
stored as a single json object instead of 2 separeta entries. This
halves database requests when looking them up.

Implemented a getter for the default_graph value that is to be accessed
via ajax.

Signed-off-by: Axel Bolle <abolle@itrsgroup.com>