Make default graph actually do something (#258)
Introduce a new behavior for the graph button in list views.
When you now hover over the graph icon in a list view a lookup will be done
to see if the setting "default_graph" is set on that particular host. If
set, it will display the corresponding graph from pnp.

Also changed the way we save default_graph in the database. It is now
stored as a single json object instead of 2 separate entries. This
halves database requests when looking them up, due to our ORM.

Implemented a getter for the default_graph value that is to be accessed
via ajax.

This solves: MON-13020
Signed-off-by: Axel Bolle <abolle@itrsgroup.com>