Implement simple caching for listview graph hover
This commit will add a counter to each hoverable element which
increments on hover.
The first hover for every page reload triggered a database lookup of
the default_graph setting. After that it will only do a lookup every 5
hovers, or until a page reload is done.

This limits the ammount of database calls and still provides adequate
real time graph visiblity.

Signed-off-by: Axel Bolle <abolle@itrsgroup.com>