Remove test_this_* permission from limited_edit
* Remove permissions test_this_host and test_this_service from the
  limited_edit group that is distributed by default on new
  installations.

* Remove from default auth_groups.yml file - affects new installations
  only.

* Add version to migrate_auth script that removes
  test_this_host/test_this_service from the group "limited_edit", from
  existing configurations. This affects existing installations,
  including when restoring an older auth_groups.yml file from a
  (migration) backup with op5-restore (as it triggers ninjas %post
  scriptlet, which in turn runs the migration script).

Background: The permissions "test this host" and "test this service"
more or less grants the user shell access as the monitor user on the
host, which could be used to intentionally or by mistake do harm. Files
can be modified, all databases accessed etc. So in reality the user is
not really "limited" to objects he/she is contact for.

Fixes MON-12535.

Signed-off-by: Aksel Sjögren <asjogren@itrsgroup.com>