Undo config after backup restore
The Backup/Restore feature in Ninja would only restore the configuration
files without touching the Nacoma database. This could result in Nacoma
not matching the actual configuration, if file changes are not detected.
With this commit we run the `undo_config` action in the Nacoma API as
part of the restore, which discards reloads the config in the database
from the config files.

This is part of MON-13039.

Signed-off-by: Petter Nyström <pnystrom@itrsgroup.com>