Add --all-nodes option to showlog
This commit adds a new command line option --all-nodes to the
showlog tool. This new option will gather logs from all nodes
and create a consolidated list of event logs. Each log is
appended with the node name where the log originated.
A new radio button is added to the event log page to invoke
the new  command line option.

This is part of MON-13186

Signed-off-by: Jerson Dumalaon <jdumalaon@itrsgroup.com>