Update http directives (#287)
In this commit we update deprecated http directives.

Note: In MON-11155 work was carried out to restrict/deny access to
source code files through http. For the Location /ninja/install_scripts
it seems as the idea was to deny access but in reality it was not.
Looking at latest versions of Monitor 9 and 8 files are still accessible.
With this commit we update the deprecated directives and continue to
grant access for that Location. We will create a new story to
investigate if we could/should deny access.

This is part of MON-13171

Signed-off-by: Daniel Nilsson <dnilsson@itrsgroup.com>