Update filter when modified
This commit checks if there is an existing filter name under the same username then update it instead of deleting and re-inserting the filter details. Global filters were saved will a NULL username.
Global filters can have the same name as local filters. Saving a filter with the same name and scope but with different filter_table will only update the filter query.

This resolves MON-13298

Signed-off-by: Eunice Remoquillo <eremoquillo@itrsgroup.com>