Update filter when modified (#308)
This commit checks if there is an existing filter name under the same username then update it using the existing id instead of deleting and re-inserting the filter details. This will make sure that the id being referenced by a dashboard filter will be kept upon updating the filter.
Global filters were saved will a NULL username. Global filters can have the same name as local filters. Saving a filter with the same name and scope but with different filter_table will only update the filter query. Users should have a Saved Filter Global permission to be able to modify a global filter.

This resolves MON-13298

Signed-off-by: Eunice Remoquillo <eremoquillo@itrsgroup.com>