Sets a default sort value for comments
A default value is set for sort variable when the query is intended for comments table. This value will only be used on initial load of the page and if the user has not selected a sorting order yet.

This should not affect other extinfo page tables that are not related to comments.

This resolves MON-10404.

Signed-off-by: Eunice Remoquillo <eremoquillo@itrsgroup.com>