Temporary fix to display graph on generated pdf
This commit is a temporary solution to be able
to display generated Trends Graph on pdf file.
Since the availability report is the only report
that uses this function, the whole TGraph() is
moved to the only file using it.

The difference from Mon8's were
ninja's head_template script and $(window).on('load') syntax.

Will do further investigations once related
functions were needed to be tested.

Signed-off-by: Eunice Remoquillo <eremoquillo@itrsgroup.com>